refactored the files and folders, added a security config
