+++ secure-service/src/main/java/basicAuthentication/BasicAuthenticationApplication.java
@@ -1,13 +1,11 @@
-package com.secure_service.example.secure_service;
+package basicAuthentication;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SecureServiceApplication {
-
+public class BasicAuthenticationApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(SecureServiceApplication.class, args);
+		SpringApplication.run(BasicAuthenticationApplication.class, args);
 	}
-
 }
